#8517 ui part of plugin ext.spanner in ext.spanner.ui moved
