+++ features/org.jkiss.dbeaver.db.ui.feature/feature.xml
@@ -30,7 +30,7 @@
 
     <plugin id="org.jkiss.dbeaver.ext.athena.ui" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
     <plugin id="org.jkiss.dbeaver.ext.bigquery.ui" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
-    <plugin id="org.jkiss.dbeaver.ext.spanner" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
+    <plugin id="org.jkiss.dbeaver.ext.spanner.ui" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
     <plugin id="org.jkiss.dbeaver.ext.snowflake.ui" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
     <plugin id="org.jkiss.dbeaver.ext.sqlite.ui" download-size="0" install-size="0" version="0.0.0" unpack="false"/>
