+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,4 @@
+Bundle-Vendor = DBeaver Corp
+Bundle-Name = DBeaver Google Spanner UI
+
+spanner.dialog.connection.header=Google Spanner Connection Settings
