+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/OSGI-INF/l10n/bundle_de.properties
@@ -1,0 +1,1 @@
+spanner.dialog.connection.header=Google Spanner Verbindungseinstellungen
