+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -1,0 +1,4 @@
+Bundle-Name = DBeaver Google UI
+Bundle-Vendor = DBeaver Corp
+
+spanner.dialog.connection.header = Google Spanner Configura\u00E7\u00F5es de conex\u00E3o
