+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/plugin.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.2"?>
+
+<plugin>
+
+    <extension point="org.jkiss.dbeaver.dataSourceView">
+        <view
+                id="org.jkiss.dbeaver.ext.spanner.dataSourceWizardView"
+                targetID="org.jkiss.dbeaver.ext.ui.newConnectionWizard"
+                dataSource="google_spanner"
+                label="%spanner.dialog.connection.header"
+                class="org.jkiss.dbeaver.ext.spanner.ui.views.SpannerConnectionPage"
+                icon="platform:/plugin/org.jkiss.dbeaver.ext.spanner/icons/spanner_icon.png">
+        </view>
+        <view
+                class="org.jkiss.dbeaver.ext.spanner.ui.views.SpannerConnectionPage"
+                icon="platform:/plugin/org.jkiss.dbeaver.ext.spanner/icons/spanner_icon.png"
+                id="org.jkiss.dbeaver.ext.spanner.dataSourceEditorView"
+                dataSource="google_spanner"
+                label="%spanner.dialog.connection.header"
+                targetID="org.jkiss.dbeaver.ext.ui.editConnectionDialog">
+        </view>
+    </extension>
+
+</plugin>
