+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/src/org/jkiss/dbeaver/ext/spanner/ui/internal/messages.properties
@@ -1,3 +1,5 @@
 label_connection=Connection
 label_private_key_path=Private key path (p12 or JSON)
 label_project=Project
+label_instance=Instance 
+label_database=Database
