+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/src/org/jkiss/dbeaver/ext/spanner/ui/internal/messages_de.properties
@@ -1,0 +1,5 @@
+label_connection=Verbindung
+label_private_key_path=Privater Schl\u00fcsselpfad (p12 oder JSON)
+label_project=Projekt
+label_instance=Instanz 
+label_database=Datenbank
