+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/src/org/jkiss/dbeaver/ext/spanner/ui/internal/messages_ru.properties
@@ -1,3 +1,5 @@
 label_connection=\u0421\u043E\u0435\u0434\u0438\u043D\u0435\u043D\u0438\u0435
 label_private_key_path=\u041F\u0443\u0442\u044C \u043A \u0437\u0430\u043A\u0440\u044B\u0442\u043E\u043C\u0443 \u043A\u043B\u044E\u0447\u0443 (p12 \u0438\u043B\u0438 JSON)
 label_project=\u041F\u0440\u043E\u0435\u043A\u0442
+label_instance=\u042d\u043a\u0437\u0435\u043c\u043f\u043b\u044f\u0440
+label_database=\u0411\u0430\u0437\u0430\u0020\u0434\u0430\u043d\u043d\u044b\u0445
