+++ plugins/org.jkiss.dbeaver.ext.spanner.ui/src/org/jkiss/dbeaver/ext/spanner/ui/views/SpannerConnectionPage.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jkiss.dbeaver.ext.spanner.views;
+package org.jkiss.dbeaver.ext.spanner.ui.views;
 
 import org.eclipse.jface.dialogs.IDialogPage;
 import org.eclipse.jface.resource.ImageDescriptor;
@@ -25,7 +25,8 @@ import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Text;
 import org.jkiss.dbeaver.Log;
-import org.jkiss.dbeaver.ext.spanner.SpannerActivator;
+import org.jkiss.dbeaver.ext.spanner.ui.SpannerActivator;
+import org.jkiss.dbeaver.ext.spanner.ui.internal.SpannerMessages;
 import org.jkiss.dbeaver.ext.spanner.model.SpannerConstants;
 import org.jkiss.dbeaver.model.DBPDataSourceContainer;
 import org.jkiss.dbeaver.model.connection.DBPConnectionConfiguration;
@@ -78,11 +79,11 @@ public class SpannerConnectionPage extends ConnectionPageAbstract implements ICo
             projectText = UIUtils.createLabelText(addrGroup, SpannerMessages.label_project, ""); //$NON-NLS-2$
             projectText.addModifyListener(textListener);
 
-            instanceText = UIUtils.createLabelText(addrGroup, "Instance", ""); //$NON-NLS-2$
+            instanceText = UIUtils.createLabelText(addrGroup, SpannerMessages.label_instance, ""); //$NON-NLS-2$
             instanceText.setToolTipText("Spanner Instance ID"); //$NON-NLS-1$
             instanceText.addModifyListener(textListener);
 
-            databaseText = UIUtils.createLabelText(addrGroup, "Database", ""); //$NON-NLS-2$
+            databaseText = UIUtils.createLabelText(addrGroup, SpannerMessages.label_database, ""); //$NON-NLS-2$
             databaseText.setToolTipText("Spanner Database ID"); //$NON-NLS-1$
             databaseText.addModifyListener(textListener);
