+++ plugins/org.jkiss.dbeaver.ext.spanner/META-INF/MANIFEST.MF
@@ -4,14 +4,10 @@ Bundle-Name: DBeaver Spanner extension
 Bundle-SymbolicName: org.jkiss.dbeaver.ext.spanner;singleton:=true
 Bundle-Version: 1.0.18.qualifier
 Bundle-Release-Date: 20200420
-Require-Bundle: org.eclipse.ui,
- org.eclipse.core.runtime,
- org.jkiss.dbeaver.core,
- org.jkiss.dbeaver.model,
- org.jkiss.dbeaver.ext.generic
-Bundle-Activator: org.jkiss.dbeaver.ext.spanner.SpannerActivator
+Require-Bundle: org.jkiss.dbeaver.ext.generic
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: DBeaver Corp
 Bundle-Localization: OSGI-INF/l10n/bundle
 Automatic-Module-Name: org.jkiss.dbeaver.ext.spanner
+Export-Package: org.jkiss.dbeaver.ext.spanner.model
