+++ plugins/org.jkiss.dbeaver.ext.spanner/OSGI-INF/l10n/bundle.properties
@@ -1,5 +1,3 @@
 Bundle-Vendor = DBeaver Corp
 Bundle-Name = DBeaver Google Spanner Extension
 datasource.spanner.description = Spanner datasource
-
-spanner.dialog.connection.header=Google Spanner Connection Settings
