+++ plugins/org.jkiss.dbeaver.ext.spanner/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -4,5 +4,3 @@ Bundle-Name = DBeaver Google Spanner Extens\u00E3o
 Bundle-Vendor = DBeaver Corp
 
 datasource.spanner.description = Spanner fonte de dados
-
-spanner.dialog.connection.header = Google Spanner Configura\u00E7\u00F5es de conex\u00E3o
